CHROMIUM: thermal: mediatek: Reland forward port of svs driver for mt8173
This the forward port of the SVS driver from chromeos-4.19. Diffs
between chromeos-4.19 are:
- Squashed Kconfig changes to this patch
- Migrated pm_qos_add_request to cpuidle_pause_and_lock
- Migrated CPUFREQ_ADJUST callback to freq_qos_add_request
- Added error handling for dev_pm_opp_adjust_voltage
- Misc style fixes

Original commit message from 4.19:

This restores the SVS portions of the mtk-thermal driver that was used
in 3.18, and uses them for mt8173. The current state is that thermal and
SVS are split into two separate drivers. However, the current SVS driver
only supports mt8183 [1]. There is no patchset for adding support for
mt8173 to the new driver, therefore this is a pragmatic approach to
restoring SVS to mt8173.

b:157105984 has been filed to track upstream SVS support for mt8173.

[1] https://lore.kernel.org/linux-mediatek/20200316072316.7156-1-roger.lu@mediatek.com/

BUG=b:157105984, b:246249061
TEST=PowerQual.full
TEST=printer.TestPPDs
UPSTREAM-TASK=b:256786714

Signed-off-by: Anand K Mistry <amistry@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2195581
Commit-Queue: Anand Mistry <amistry@chromium.org>
Tested-by: Anand Mistry <amistry@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Signed-off-by: Pin-yen Lin <treapking@chromium.org>
Change-Id: I5e38095eee1fdc0ca6ff2f3fe03ee2206237aff8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4520703
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>