UPSTREAM: PM / devfreq: Add cpu based scaling support to passive governor
Many CPU architectures have caches that can scale independent of the
CPUs. Frequency scaling of the caches is necessary to make sure that the
cache is not a performance bottleneck that leads to poor performance and
power. The same idea applies for RAM/DDR.

To achieve this, this patch adds support for cpu based scaling to the
passive governor. This is accomplished by taking the current frequency
of each CPU frequency domain and then adjust the frequency of the cache
(or any devfreq device) based on the frequency of the CPUs. It listens
to CPU frequency transition notifiers to keep itself up to date on the
current CPU frequency.

To decide the frequency of the device, the governor does one of the
following:
* Derives the optimal devfreq device opp from required-opps property of
  the parent cpu opp_table.

* Scales the device frequency in proportion to the CPU frequency. So, if
  the CPUs are running at their max frequency, the device runs at its
  max frequency. If the CPUs are running at their min frequency, the
  device runs at its min frequency. It is interpolated for frequencies
  in between.

Tested-by: Chen-Yu Tsai <wenst@chromium.org>
Tested-by: Johnson Wang <johnson.wang@mediatek.com>
Signed-off-by: Saravana Kannan <skannan@codeaurora.org>
[Sibi: Integrated cpu-freqmap governor into passive_governor]
Signed-off-by: Sibi Sankar <sibis@codeaurora.org>
[Chanwoo: Fix conflict with latest code and cleanup code]
Signed-off-by: Chanwoo Choi <cw00.choi@samsung.com>

(cherry picked from commit a03dacb0316f74400846aaf144d6c73f4217ca08)

BUG=b:213000795
TEST=build

Signed-off-by: Hsin-Yi Wang <hsinyi@chromium.org>
Change-Id: I6d3a50c8e992e40e5111684dabfca6ec3140943b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3706559
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>