FROMLIST: devfreq: add mediatek cci devfreq
This adds a devfreq driver for the Cache Coherent Interconnect (CCI)
of the Mediatek MT8183.

On the MT8183 the CCI is supplied by the same regulator as the LITTLE
cores. The driver is notified when the regulator voltage changes
(driven by cpufreq) and adjusts the CCI frequency to the maximum
possible value.

Signed-off-by: Andrew-sh.Cheng <andrew-sh.cheng@mediatek.com>

(am from https://patchwork.kernel.org/project/linux-mediatek/patch/1616499241-4906-5-git-send-email-andrew-sh.cheng@mediatek.com/)

BUG=b:129039370
TEST=boot krane

Signed-off-by: Hsin-Yi Wang <hsinyi@chromium.org>
Change-Id: I184831f7115e43d543da1a8fa276d125c428ce5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2635417
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>