UPSTREAM: arm64: dts: qcom: apq8016-sbc: Update modem and WiFi firmware path
The firmware for the modem and WiFi subsystems platform specific and is
signed with a OEM specific key (or a test key). In order to support more
than a single device it is therefor not possible to rely on the default
path and stash these files directly in the firmware directory.

This has already been addressed for other platforms, but the APQ8016 SBC
(aka db410c) was never finished upstream.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Reviewed-by: Stephan Gerhold <stephan@gerhold.net>
Tested-by: Stephan Gerhold <stephan@gerhold.net>
Link: https://lore.kernel.org/r/20210531224453.783218-1-bjorn.andersson@linaro.org
(cherry picked from commit 0f6b380d580cd081d5e385d349f55dfc52e3d68c)

BUG=none
TEST=future patches are easier to pick

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I1056ba7b6728b3b06e2a4ddac6ce0fa45bde5c8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3930983
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Douglas Anderson <dianders@chromium.org>