BACKPORT: media: uvcvideo: refactor __uvc_ctrl_add_mapping
Pass the chain instead of the device. We want to keep the reference to
the chain that controls belong to.

We need to delay the initialization of the controls after the chains
have been initialized.

This is a cleanup needed for the next patches.

Reviewed-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>

(cherry picked from commit 866c6bdd5663d4df7cf384b381b6ef8ba9ffd0e4)

Conflicts:
   drivers/media/usb/uvc/uvc_ctrl.c
   [Conflicts with CHROMIUM: roi patches]

BUG=b:182408307
TEST=v4l2-compliance

Change-Id: I77aa8090526c4afe814ed154bce5c2803464055f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3282112
Commit-Queue: Ricardo Ribalda <ribalda@chromium.org>
Tested-by: Ricardo Ribalda <ribalda@chromium.org>
Reviewed-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3311277
Reviewed-by: Douglas Anderson <dianders@chromium.org>