UPSTREAM: PCI: dwc: Perform host_init() before registering msi
On the Qualcomm sc8180x platform the bootloader does something related
to PCI that leaves a pending "msi" interrupt, which with the current
ordering often fires before init has a chance to enable the clocks that
are necessary for the interrupt handler to access the hardware.

Move the host_init() call before the registration of the "msi" interrupt
handler to ensure the host driver has a chance to enable the clocks.

The assignment of the bridge's ops and child_ops is moved along, because
at least the TI Keystone driver overwrites these in its host_init
callback.

Link: https://lore.kernel.org/r/20210823154958.305677-1-bjorn.andersson@linaro.org
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Reviewed-by: Rob Herring <robh@kernel.org>
(cherry picked from commit 7e919677bb392c765e0f86fda63191e517c0bd1f)

BUG=b:217106877
TEST=NVMe still probes on herobrine

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I7df99936061a57bc272ab25b4ca3b09e068fd94b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3727266
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>