UPSTREAM: sched/topology: Introduce sched_group::flags
There exist situations in which the load balance needs to know the
properties of the CPUs in a scheduling group. When using asymmetric
packing, for instance, the load balancer needs to know not only the
state of dst_cpu but also of its SMT siblings, if any.

Use the flags of the child scheduling domains to initialize scheduling
group flags. This will reflect the properties of the CPUs in the
group.

A subsequent changeset will make use of these new flags. No functional
changes are introduced.

Originally-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Reviewed-by: Joel Fernandes (Google) <joel@joelfernandes.org>
Reviewed-by: Len Brown <len.brown@intel.com>
Reviewed-by: Vincent Guittot <vincent.guittot@linaro.org>
Link: https://lkml.kernel.org/r/20210911011819.12184-3-ricardo.neri-calderon@linux.intel.com
(cherry picked from commit 16d364ba6ef2aa59b409df70682770f3ed23f7c0)

BUG=b:179699891
TEST=basic boot on brya and check /proc/sys/kernel/sched_itmt_enabled
     should be 1
TEST=spin N+1 threads where N = number of physical cores - expect
        N threads to run on N physical cores and the N+1st thread
        to run on an atom CPU

Change-Id: Icba999d4baf944680d581d9707c04dedd333f76f
Disallow-Recycled-Builds: test-failures
Signed-off-by: Guenter Roeck <groeck@google.com>