UPSTREAM: drm/virtio: implement context init: add virtio_gpu_fence_event
Similar to DRM_VMW_EVENT_FENCE_SIGNALED.  Sends a pollable event
to the DRM file descriptor when a fence on a specific ring is
signaled.

One difference is the event is not exposed via the UAPI -- this is
because host responses are on a shared memory buffer of type
BLOB_MEM_GUEST [this is the common way to receive responses with
virtgpu].  As such, there is no context specific read(..)
implementation either -- just a poll(..) implementation.

Signed-off-by: Gurchetan Singh <gurchetansingh@chromium.org>
Acked-by: Nicholas Verne <nverne@chromium.org>
Link: http://patchwork.freedesktop.org/patch/msgid/20210921232024.817-12-gurchetansingh@chromium.org
Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>
(cherry picked from commit cd7f5ca33585918febe5e2f6dc090a21cfa775b0)

BUG=b:216688312
TEST=glxgears and vkcube in a crosvm guest.

Signed-off-by: Ryan Neph <ryanneph@google.com>
Change-Id: I3806a46b1524f21787e5c9f3f75367e26849ca4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3425177
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>