UPSTREAM: drm/i915/gem: Implement a function to process all gem objects of a region
An upcoming common pattern is to traverse the region object list and
perform certain actions on all objects in a region. It's a little tricky
to get the list locking right, in particular since a gem object may
change region unless it's pinned or the object lock is held.

Define a function that does this for us and that takes an argument that
defines the action to be performed on each object.

v3:
- Improve structure documentation a bit (Matthew Auld)

Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>
Reviewed-by: Matthew Auld <matthew.auld@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20210922062527.865433-3-thomas.hellstrom@linux.intel.com
(cherry picked from commit d80ee88e0769e2e05afeb5d04b4dc43fc107b0d5)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:214079760
TEST=None

Backported using forklift.py

Change-Id: I66302d06f2f28d91054b33a99c1f4d87d771977f