UPSTREAM: vfio: Move vfio_iommu_group_get() to vfio_register_group_dev()
We don't need to hold a reference to the group in the driver as well as
obtain a reference to the same group as the first thing
vfio_register_group_dev() does.

Since the drivers never use the group move this all into the core code.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>
Signed-off-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Kevin Tian <kevin.tian@intel.com>
Link: https://lore.kernel.org/r/20210924155705.4258-2-hch@lst.de
Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
(cherry picked from commit 38a68934aa72459217986cf6b461d87f7602648a)

BUG=b:242065767
TEST=build test

Signed-off-by: Drew Davenport <ddavenport@chromium.org>
Change-Id: I1a361186a6cb65edb2390a0a01b876dbd126415b