UPSTREAM: drm/msm/dp: Remove redundant initialization of variable bpp
The variable bpp is being initialized with a value that is never
read, it is being updated later on in both paths of an if statement.
The assignment is redundant and can be removed.

Addresses-Coverity: ("Unused value")
Signed-off-by: Colin Ian King <colin.king@canonical.com>
Reviewed-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>
Link: https://lore.kernel.org/r/20210929105458.209895-1-colin.king@canonical.com
Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>
Signed-off-by: Rob Clark <robdclark@chromium.org>
(cherry picked from commit 76544e4bb1a05818181bc36d093a561f287afeba)

BUG=b:201590452, b:177187240
TEST=emerge-herobrine chromeos-kernel-5_10

Change-Id: I42f31c383ec5d3d4493862392ddc23279e3c3b02
Disallow-Recycled-Builds: test-failures
Signed-off-by: Guenter Roeck <groeck@google.com>