UPSTREAM: drm/mediatek: mtk_dsi: Reset the dsi0 hardware
Reset dsi0 HW to default when power on. This prevents to have different
settingis between the bootloader and the kernel.

As not all Mediatek boards have the reset consumer configured in their
board description, also is not needed on all of them, the reset is optional,
so the change is compatible with all boards.

Cc: Jitao Shi <jitao.shi@mediatek.com>
Suggested-by: Chun-Kuang Hu <chunkuang.hu@kernel.org>
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Acked-by: Chun-Kuang Hu <chunkuang.hu@kernel.org>
Reviewed-by: Matthias Brugger <matthias.bgg@gmail.com>
Link: https://lore.kernel.org/r/20210930103105.v4.7.Idbb4727ddf00ba2fe796b630906baff10d994d89@changeid
Signed-off-by: Matthias Brugger <matthias.bgg@gmail.com>
(cherry picked from commit 605c83753d97946aab176735020a33ebfb0b4615)

BUG=b:213846857
TEST=emerge-corsola sys-kernel/chromeos-kernel-5_15

Signed-off-by: Rex-BC Chen <rex-bc.chen@mediatek.corp-partner.google.com>
Change-Id: Idbb4727ddf00ba2fe796b630906baff10d994d89
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3392830
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Commit-Queue: Hsin-Yi Wang <hsinyi@chromium.org>
Tested-by: Hsin-Yi Wang <hsinyi@chromium.org>