UPSTREAM: dt-bindings: display: simple: Add Vivax TPC-9150 panel
This tablet display is unique and hence must be added to documentation.
Unfortunately, it's real make and number is not known due to obscure and
missing documentation and lack of writing on the display assembly
itself.

So the natural alternative that came to mind is to name the panel after
the device that contained it (an Android tablet). Hopefully that isn't
an issue.

I can confirm that the parameters that I've obtained are correct since
they came from a dump of the original Android OS.

Signed-off-by: Nikola Pavlica <pavlica.nikola@gmail.com>
Signed-off-by: Sam Ravnborg <sam@ravnborg.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20211002202905.509437-2-pavlica.nikola@gmail.com
(cherry picked from commit f474bb3000b6114b8beea3170d9cbed963593467)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:214079760
TEST=None

Backported using forklift.py

Change-Id: I82bc1f377e6c6bd98e7caa6a32fefce2126d5deb