UPSTREAM: platform/x86: thinkpad_acpi: Register a privacy-screen device
Register a privacy-screen device on laptops with a privacy-screen,
this exports the PrivacyGuard features to user-space using a
standardized vendor-agnostic sysfs interface. Note the sysfs interface
is read-only.

Registering a privacy-screen device with the new privacy-screen class
code will also allow the GPU driver to get a handle to it and export
the privacy-screen setting as a property on the DRM connector object
for the LCD panel. This DRM connector property is a new standardized
interface which all user-space code should use to query and control
the privacy-screen.

Reviewed-by: Emil Velikov <emil.l.velikov@gmail.com>
Reviewed-by: Lyude Paul <lyude@redhat.com>
Reviewed-by: Mark Pearson <markpearson@lenovo.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20211005202322.700909-9-hdegoede@redhat.com
(cherry picked from commit f809891ee51b706e1a2a42998d8766c120660796)

BUG=b:196060866
TEST=Build & Boot

Signed-off-by: Rajat Jain <rajatja@google.com>
Change-Id: Ic42ac197bb28ec99fa9a4876b1c331ea94590890
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3380373
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>