UPSTREAM: platform/x86: amd-pmc: adjust arguments for `amd_pmc_send_cmd`
Currently the "argument" for the "message" is listed as a boolean
value.  This works well for the commands used currently, but an
additional upcoming command will pass more data in the message.

Expand it to be a full 32 bit value.

Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>
Link: https://lore.kernel.org/r/20211020162946.10537-1-mario.limonciello@amd.com
Reviewed-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
(cherry picked from commit 4c9dbf8622797eee44e9a4f7327d3d04aa4d1d5f)

BUG=b:263448869
TEST=emerge-skyrim-chausie chromeos-kernel-5_15

Change-Id: Iffd32858ae79a35e8974145919be0f451faa492e
Signed-off-by: Harsh Jain <harsh.jain@amd.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3750782
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>