UPSTREAM: ath11k: Increment pending_mgmt_tx count before tx send invoke
There is a race condition whereby the tx completion handler can be invoked
before the 'num_pending_mgmt_tx" count is incremented. If that occurs, we
could get warning trace indicating that 'num_pending_mgmt_tx' is 0 (because
it was not yet incremented). Ideally, this trace should be seen only if
mgmt tx has not happened but tx completion is received, and it is not
expected in this race condition.

Tested-on: QCN9074 hw1.0 PCI WLAN.HK.2.4.0.1-01386-QCAHKSWPL_SILICONZ-1

Co-developed-by: Lavanya Suresh <lavaks@codeaurora.org>
Signed-off-by: Lavanya Suresh <lavaks@codeaurora.org>
Signed-off-by: P Praneesh <quic_ppranees@quicinc.com>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
Link: https://lore.kernel.org/r/1635168282-8845-1-git-send-email-quic_ppranees@quicinc.com
(cherry picked from commit c0b0d2e87d91ce283c8766b4b3c2ec9ac90ebf96)

BUG=b:232096343
TEST=emerge-herobrine chromeos-kernel-5_15; deploy and test basic WiFi
     functionality on herobrine device

Signed-off-by: Kevin Lund <kglund@google.com>
Change-Id: I13507005c7e0de26dc718e12e4c613be45654724
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3653060
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>