UPSTREAM: drm/vc4: hdmi: Check the device state in prepare()
Even though we already check that the encoder->crtc pointer is there
during in startup(), which is part of the open() path in ASoC, nothing
guarantees that our encoder state won't change between the time when we
open the device and the time we prepare it.

Move the sanity checks we do in startup() to a helper and call it from
prepare().

Link: https://lore.kernel.org/r/20211025141113.702757-8-maxime@cerno.tech
Fixes: 91e99e113929 ("drm/vc4: hdmi: Register HDMI codec")
Acked-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
(cherry picked from commit a64ff88cb5eb0b9a6855a24ff326e948931e3a8e)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:227338908
TEST=build test

Backported using forklift.py

Change-Id: Iff78754cc00420e73ae643f638212532af5586df