UPSTREAM: dma-direct: fail allocations that can't be made coherent
If the architecture can't remap or set an address uncached there is no way
to fullfill a request for a coherent allocation.  Return NULL in that case.
Note that this case currently does not happen, so this is a theoretical
fixup and/or a preparation for eventually supporting platforms that
can't support coherent allocations with the generic code.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Robin Murphy <robin.murphy@arm.com>
(cherry picked from commit 955f58f7406ad912825fc344c7825fd904b124a0)

BUG=b:259163250
TEST=Test Audio use cases

Change-Id: I5edab172745d26b5cace0ec13227431d0d87959a
Signed-off-by: Vamshi Krishna Gopal <vamshi.krishna.gopal@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4274032
Reviewed-by: Terry Cheong <htcheong@chromium.org>
Tested-by: Terry Cheong <htcheong@chromium.org>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>