UPSTREAM: dt-bindings: mtd: spi-nor: Add a reference to spi-peripheral-props.yaml
The spi-peripheral-props.yaml schema contains peripheral-specific
properties for SPI controllers that should be present in the peripheral
node. Add a reference to that so its constraints are followed.

additionalProperties: false cannot be used since it marks the controller
properties as unknown. Use unevaluatedProperties: false instead. This
has the side effect of allowing extra properties that are not specified
in the schema. The alternative is to list all the controller properties
in this schema but that would mean every peripheral binding would have
to repeat the same set of properties for each controller.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Reviewed-by: Rob Herring <robh@kernel.org>
Link: https://lore.kernel.org/r/20211109181911.2251-4-p.yadav@ti.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit e9d7c323cfbbd07c365a419b4ce3dc2f161442c7)

BUG=b:179533783
TEST=make dtbs_check

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: Ic20007e588f2126ec532670f24888edc77a22c6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4018255
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Douglas Anderson <dianders@chromium.org>