UPSTREAM: media: hantro: Avoid global variable for jpeg quantization tables
On a system with multiple encoders present, it's possible for two
encoders to write to the global luma and chroma quantization tables at
the same time if they both submit a JPEG frame to be encoded. Avoid
this race by moving the tables into the `jpeg_ctx` structure which is
stored on the stack.

Signed-off-by: James Cowgill <james.cowgill@blaize.com>
Reviewed-by: Ezequiel Garcia <ezequiel@vanguardiasur.com.ar>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit 41479adb5e52998ecf690b46164fffea3f5d65fd)

BUG=b:261593922
TEST=emerge-arm64-generic chromeos-kernel-5_15

Change-Id: I3f3f70d1a728297eb755f63e99cd32590a66e0fb
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4328502
Reviewed-by: Steve Cho <stevecho@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>